requirements.txt
appnope==0.1.0
backcall==0.2.0
beautifulsoup4==4.9.3
boto3==1.16.57
botocore==1.19.57
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
decorator==4.4.2
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
jedi==0.17.2
jmespath==0.10.0
jupyter-client==6.1.7
jupyter-core==4.6.3
mysql-connector-python==8.0.23
numpy==1.20.0
pandas==1.2.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.8
protobuf==3.14.0
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
pyzmq==19.0.2
requests==2.25.1
s3transfer==0.3.4
six==1.15.0
soupsieve==2.1
tornado==6.0.4
tqdm==4.56.2
traitlets==5.0.5
urllib3==1.26.2
wcwidth==0.2.5

fsspec~=0.8.7
setuptools~=54.0.0
